deps: drop dependency on ts-amqp
fixes #19
