+++ package.json
@@ -55,7 +55,6 @@
   "dependencies": {
     "amqplib": "^0.10.7",
     "debug": "^4.4.0",
-    "ts-amqp": "^0.8.1",
     "uuid": "^9.0.0"
   }
 }
