+++ src/app.ts
@@ -6,82 +6,19 @@ import { PeanarInternalError } from './exceptions';
 import Broker from './broker';
 import Worker, { IWorkerResult } from './worker';
 import PeanarJob from './job';
-import { IConnectionParams } from 'ts-amqp/dist/interfaces/Connection';
 import { Writable, TransformCallback } from 'stream';
-import { IBasicProperties } from 'ts-amqp/dist/interfaces/Protocol';
 import Registry from './registry';
 import Consumer from './consumer';
 import { Channel } from 'amqplib';
-
-export interface IPeanarJobDefinitionInput {
-  queue: string;
-  name?: string;
-  routingKey?: string;
-  exchange?: string;
-  handler: (...args: any[]) => Promise<any>;
-
-  jobClass?: typeof PeanarJob;
-
-  expires?: number;
-  retry_exchange?: string;
-  error_exchange?: string;
-  max_retries?: number;
-  retry_delay?: number;
-  delayed_run_wait?: number;
-
-  max_priority?: number;
-  default_priority?: number;
-}
-
-export interface IPeanarJobDefinition {
-  name: string;
-  queue: string;
-  handler: (...args: any[]) => Promise<any>;
-
-  routingKey: string;
-  exchange?: string;
-
-  jobClass?: typeof PeanarJob;
-
-  expires?: number;
-  retry_exchange?: string;
-  error_exchange?: string;
-  max_retries?: number;
-  retry_delay?: number;
-  delayed_run_wait?: number;
-
-  max_priority?: number;
-  default_priority?: number;
-}
-
-export interface IPeanarRequest {
-  id: string;
-  name: string;
-  args: any[];
-  attempt: number;
-  deliveryTag?: bigint;
-  priority?: number;
-}
-
-export interface IPeanarJob extends IPeanarJobDefinition, IPeanarRequest {
-  deliveryTag: bigint;
-}
-
-export interface IPeanarResponse {
-  id: string;
-  name: string;
-  status: 'SUCCESS' | 'FAILURE';
-  error?: unknown;
-  result?: unknown;
-}
-
-export interface IPeanarOptions {
-  connection?: IConnectionParams;
-  poolSize?: number;
-  prefetch?: number;
-  jobClass?: typeof PeanarJob;
-  logger?(...args: any[]): any;
-}
+import {
+  EAppState,
+  IPeanarOptions,
+  IPeanarJobDefinition,
+  IPeanarRequest,
+  IBasicProperties,
+  IPeanarResponse,
+  IPeanarJobDefinitionInput
+} from './types';
 
 interface IWorkerOptions {
   queues?: string[];
@@ -89,12 +26,6 @@ interface IWorkerOptions {
   prefetch?: number;
   outputWritable?: Writable;
   logger?: (msg: string) => void;
-}
-
-export enum EAppState {
-  RUNNING = 'RUNNING',
-  CLOSING = 'CLOSING',
-  CLOSED = 'CLOSED'
 }
 
 export default class PeanarApp {
