+++ src/broker.ts
@@ -5,43 +5,19 @@ import amqplib, { ChannelModel, ConsumeMessage, Replies, Channel } from 'amqplib
 
 import { ChannelPool } from './pool';
 import { PeanarAdapterError } from './exceptions';
+import {
+  IBasicProperties,
+  IBinding,
+  IConnectionParams,
+  IExchange,
+  IQueue
+} from './types';
 import Consumer from './consumer';
-
-interface IConnectionParams {
-  maxRetries: number;
-  retryDelay: number;
-  host: string;
-  port: number;
-  username: string;
-  password: string;
-  locale: string;
-  keepAlive?: boolean;
-  keepAliveDelay?: number;
-  timeout?: number;
-  vhost: string;
-}
 
 interface IBrokerOptions {
   connection?: IConnectionParams;
   poolSize: number;
   prefetch?: number;
-}
-
-interface IBasicProperties {
-  contentType?: string;
-  contentEncoding?: string;
-  headers?: Record<string, unknown>;
-  deliveryMode?: number;
-  priority?: number;
-  correlationId?: string;
-  replyTo?: string;
-  expiration?: string;
-  messageId?: string;
-  timestamp?: Date;
-  type?: string;
-  userId?: string;
-  appId?: string;
-  clusterId?: string;
 }
 
 interface IMessage<B = Buffer> {
@@ -53,41 +29,6 @@ interface IMessage<B = Buffer> {
   body?: B;
 }
 
-type EExchangeType = 'direct' | 'fanout' | 'topic' | 'headers';
-interface IExchangeArgs {
-  alternameExchange?: string;
-}
-interface IExchange {
-  name: string;
-  type: EExchangeType;
-  durable: boolean;
-  arguments?: IExchangeArgs;
-}
-
-interface IQueueArgs {
-  deadLetterExchange?: string;
-  deadLetterRoutingKey?: string;
-  expires?: number;
-  lazy?: boolean;
-  maxLength?: number;
-  maxLengthBytes?: number;
-  maxPriority?: number;
-  messageTtl?: number;
-  overflow?: 'drop-head' | 'reject-publish';
-  queueMasterLocator?: boolean;
-}
-interface IQueue {
-  name: string;
-  durable: boolean;
-  exclusive: boolean;
-  auto_delete: boolean;
-  arguments?: IQueueArgs;
-}
-interface IBinding {
-  queue: string;
-  exchange: string;
-  routing_key: string;
-}
 function timeout(ms: number) {
   return new Promise(res => {
     setTimeout(res, ms)
